# [VL] Gluten-it: Improve test report table rendering