--- a/tools/gluten-it/common/src/main/scala/org/apache/gluten/integration/action/QueriesCompare.scala
+++ b/tools/gluten-it/common/src/main/scala/org/apache/gluten/integration/action/QueriesCompare.scala
@@ -130,7 +130,7 @@ object QueriesCompare {
           "%s -> %s".format(
             line.expectedExecutionTimeMillis.getOrElse("N/A"),
             line.actualExecutionTimeMillis.getOrElse("N/A")),
-          speedUp.map("%+.2f%%".format(_)).getOrElse("N/A"))
+          speedUp.map("%.2f%%".format(_)).getOrElse("N/A"))
       }
     }
   }