--- a/tools/gluten-it/common/src/main/scala/org/apache/gluten/integration/action/TableRender.scala
+++ b/tools/gluten-it/common/src/main/scala/org/apache/gluten/integration/action/TableRender.scala
@@ -164,6 +164,17 @@ object TableRender {
           }
       }
 
+      val separationLine = schema.leafs
+        .map { leaf =>
+          widthMap(System.identityHashCode(leaf))
+        }
+        .map { width =>
+          new String(Array.tabulate(width)(_ => '-'))
+        }
+        .mkString("|", "|", "|")
+
+      printer.println(separationLine)
+
       data.foreach { row =>
         val dataLine = row
           .zip(schema.leafs)