… tools/gluten-it/common/src/main/scala/org/apache/gluten/integration/action/TableFormatter.scala
… tools/gluten-it/common/src/test/java/org/apache/gluten/integration/action/TableRenderTest.scala